+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## v1.3.0
 * fix: `ContextMenu.create` option passed `MenuOption.width` not work.
 * add: `MenutOption.title/tip` can pass formatter function.
+* add: `MenutOption.icon`
+* add: `MenuOption.customItem` 用户自定义渲染
 ## v1.2.2
 * MenuWrapper 支持传入position，不一定要MouseEvent
 * `ContextMenuOption` 支持传入默认width